test: split ApiRateLimitRequestResolver limiter tests to satisfy phpmd method limit
Move sign-in, password-reset, format-extension, and GraphQL limiter test
methods out of ApiRateLimitRequestResolverLimitersTest into a new
ApiRateLimitRequestResolverSignInLimitersTest so each class stays under the
phpmd TooManyMethods threshold (25). No assertions changed or removed.
